requirements.txt
# Required Python packages
# Usage: pip install -r requirements.txt

# Basic packages
numpy>=1.21.0
pandas>=1.3.0
requests>=2.26.0

# Web Framework
flask>=2.0.0
flask-cors>=4.0.0
django>=3.2.0

# Database
sqlalchemy>=1.4.0
pymongo>=3.12.0

# Data Visualization
matplotlib>=3.4.0
seaborn>=0.11.0

# Machine Learning
scikit-learn>=0.24.0
tensorflow>=2.6.0

# Utility packages
python-dotenv>=0.19.0
pyyaml>=5.4.0
python-decouple>=3.8.0
